Improve encapsulation of NameService artifacts.
Move Core and POAServant classes for NamingContext and BindingIterator
to be nested (static) classes under their LocalObject variants, which
are renamed to be simply NamingContextImpl and BindingIteratorImpl.

NB: The setup using post_init() in the example TransientNameService code
is still not right, I believe - it's too soon to be activating POAs.